small refacto
